Move the serialize character script to the redis root folder since there aren't any other deserialize functions
